+++ src/pages/Main/Main.scss
@@ -11,7 +11,7 @@
     .mainSlogan {
       width: 80%;
       height: 500px;
-      margin: 80px auto;
+      margin: 120px auto;
 
       img {
         width: 100%;
@@ -141,11 +141,11 @@
   }
   .mainMsg {
     width: 100%;
-    margin-top: 50px;
+    margin-top: 100px;
 
     img {
       display: block;
-      width: 80%;
+      width: 70%;
       margin: 0 auto;
     }
   }
