--- src/pages/UserAccount/Login/Login.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import UserAccount from '../UserAccount';
-import './Login.scss';
-
-function Login() {
-  return (
-    <div className="login">
-      <UserAccount />
-    </div>
-  );
-}
-
-export default Login;
