--- src/pages/UserAccount/Login/Login.scss
@@ -1,51 +1,0 @@
-.login {
-  width: 100vw;
-
-  .loginInner {
-    width: 460px;
-    margin: 0 auto;
-
-    .logo {
-      height: 50px;
-
-      img {
-        height: 60px;
-      }
-    }
-    h2 {
-      margin: 20px 0;
-      font-size: 2.8rem;
-      font-weight: 300;
-    }
-
-    .loginInput {
-      margin-top: 50px;
-
-      input {
-        width: 100%;
-        height: 48px;
-        margin-bottom: 20px;
-        padding: 13px;
-        border: 1px solid #111;
-        border-radius: 10px;
-        font-size: 1.6rem;
-      }
-    }
-    .serviceTerms {
-      width: 85%;
-      margin-top: 20px;
-      margin-bottom: 30px;
-      font-size: 1.6rem;
-      font-weight: 300;
-      line-height: 24px;
-    }
-    .loginBtn {
-      width: 100%;
-      height: 60px;
-      border-radius: 50px;
-      background-color: #111;
-      color: #fff;
-      font-size: 1.6rem;
-    }
-  }
-}
