--- src/pages/UserAccount/SignUp/SignUp.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import UserAccount from '../UserAccount';
-import '../SignUp.scss';
-
-function SignUp() {
-  return (
-    <div className="SignUp">
-      <UserAccount />
-    </div>
-  );
-}
-
-export default SignUp;
